Fix to include labels
CLIPForImageClassification isn't want w ewant.  It isn't fine tuned to so it Label_0, Label_1 etc.

From logs:
"Some weights of CLIPForImageClassification were not initialized from the model checkpoint at openai/clip-vit-base-patch32 and are newly initialized: ['classifier.bias', 'classifier.weight']
You should probably TRAIN this model on a down-stream task to be able to use it for predictions and inference."