Task/add CLIP and other models (#6)
* Rename MODEL_NAME to DEFAULT_MODEL_NAME

* Add CLIP. not working but draft

* Fix to include labels

CLIPForImageClassification isn't want w ewant.  It isn't fine tuned to so it Label_0, Label_1 etc.

From logs:
"Some weights of CLIPForImageClassification were not initialized from the model checkpoint at openai/clip-vit-base-patch32 and are newly initialized: ['classifier.bias', 'classifier.weight']
You should probably TRAIN this model on a down-stream task to be able to use it for predictions and inference."

* Add previous code

* Add way to preload models

* Add additional vit models and microsoft/swin

* Add aggregation approach

* Add missing preload_models.py

* Fix clip

* Add metadata and fix CLIP

* Fix linting errors

* Fix tests

* Refactor into multiple files

* Add test and notes for image location

* Fix geolocation metadata flow

* Drop exifread as using only Pillow now